requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
pyyaml>=6.0
gspread>=6.0.0
google-auth>=2.0.0
openpyxl>=3.1.0
